--- pkg/google/compute/testdata/.gitignore
@@ -1,1 +1,0 @@
-*.json
