+++ pkg/google/gke/machinespec_test.go
@@ -1,4 +1,4 @@
-package compute
+package gke
 
 import (
 	"testing"
