+++ pkg/google/gke/pricing_map.go
@@ -1,4 +1,4 @@
-package compute
+package gke
 
 import (
 	"context"
